Add transaction flow (#21)
* Adding the transaction flow diagram

As this may be useful to customers to visualise what is being
documented. I have also included the api-flow-diagram which is the xml
file that generated the png using draw.io incase we want to update it
and regenerate the image later.

* Adding some setup and running information

This is just so that people know how to setup and run the application
locally so they can test their changes before creating a pull request.