Adds documentation for postmark
It is often requested how to prototype or test our API without having to
use SDKs or programming implementations. Due to the nature of our
authentication mechanism this is complex to explain. As a result we have
a working postman collection that can be given to customers to help
reduce the complexity of sending basic requests to our API.