Add postman documentation (#179)
* Update to ruby 3.0

This updates the ruby version being used to a more recent (and
supported) version. This also updates the necessary libraries and
dependencies to support the ruby version update.

* Adds documentation for postmark

It is often requested how to prototype or test our API without having to
use SDKs or programming implementations. Due to the nature of our
authentication mechanism this is complex to explain. As a result we have
a working postman collection that can be given to customers to help
reduce the complexity of sending basic requests to our API.

* Fixes the search bar

It was found in local testing that the move from transferzero.com to
azafinance.com for the website has broken the link to Google search.
This commit updates this and resolves the issue that search results were
returning no results.